Provide Record(s) as Serialize Method Arg
Move the `record` (e.g. the active record instance, query) input to the
`serialize` and `serialize_page` methods. Move the `skip_includes`
setting to the initialize method. This is preferable, because it's more
useful to have a reusable serializer instance with some configured
settings than it is to have a reusable serializer instance for a
particular `record`.